+++ docs/src/docs/SankeyDiagram/propDocs.json
@@ -1,5 +1,5 @@
 {
-  "description": "A Sankey diagram is a type of flow diagram which represents",
+  "description": "A Sankey diagram is a type of flow diagram which visualizes directed flow between nodes\nof a network graph. Currently only *acyclic* networks are supported.",
   "methods": [
     {
       "name": "_makeSankeyGraph",
