+++ src/util.js
@@ -10,3 +10,14 @@ export function methodIfFuncProp(propName, props, context) {
 export function hasOneOfTwo(a, b) {
     return _.some([a, b], _.isUndefined) && _.some([a, b], v => !_.isUndefined(v));
 }
+
+function componentName(Component) {
+  return Component.displayName || "Component wrapped by resolveObjectProps";
+}
+
+function hasSome(obj, keys) {
+  return _.isObject(obj) && _.some(keys, k => _.has(obj, k));
+}
+function hasAll(obj, keys) {
+  return _.isObject(obj) && _.every(keys, k => _.has(obj, k));
+}
